### 2023-07-08 #5632 @editorialbot
[PRE REVIEW]: BoxKit: A Python library to manage analysis of block-structured simulation datasets
**Submitting author:** @akashdhruv (<a href="http://orcid.org/0000-0003-4997-321X">Akash Dhruv</a>)
**Repository:** https://github.com/akashdhruv/BoxKit
**Branch with paper.md** (empty if default branch): main
**Version:** 2023.06
