### 2023-07-08 #5632 @gkthiruvathukal
@openjournals/joss-eics I have opened this up for pre-review. It clearly belongs to my track but I have a conflict. The author is from Argonne National Laboratory. This presents an *institutional conflict* for me as I hold an appointment there.  Although I do not know or interact with the author, I feel conflicted nevertheless. Can someone please act as the AEIC for this submission instead of me? I'll also reach out on our Slack. Ideally, I think @arfon or @danielskatz can help here, given the subject matter.
