+++ paper/paper.bib
@@ -5,16 +5,19 @@
 	Pages = {101168},
 	Year = {2022},
 	issn = {2352-7110},
-	doi = {https://doi.org/10.1016/j.softx.2022.101168},
+	doi = {10.1016/j.softx.2022.101168},
 	url = {https://www.sciencedirect.com/science/article/pii/S2352711022001030},
 	Author = {Anshu Dubey and Klaus Weide and Jared O’Neal and Akash Dhruv and Sean Couch and J. Austin Harris and Tom Klosterman and Rajeev Jain and Johann Rudi and Bronson Messer and Michael Pajkos and Jared Carlson and Ran Chu and Mohamed Wahib and Saurabh Chawdhary and Paul M. Ricker and Dongwook Lee and Katie Antypas and Katherine M. Riley and Christopher Daley and Murali Ganapathy and Francis X. Timmes and Dean M. Townsley and Marcos Vanella and John Bachan and Paul M. Rich and Shravan Kumar and Eirik Endeve and W. Raphael Hix and Anthony Mezzacappa and Thomas Papatheodore}
 }
 
-@article{DHRUV2023,
-	Author = {Akash Dhruv},
-	Title = {A Vortex Damping Outflow Forcing for Multiphase Flows with Sharp Interfacial Jumps},
-	Year = {2023},
-	Journal = {Under Review}
+@misc{DHRUV2023,
+      title={A Vortex Damping Outflow Forcing for Multiphase Flows with Sharp Interfacial Jumps}, 
+      author={Akash Dhruv},
+      year={2023},
+      eprint={2306.10174},
+      archivePrefix={arXiv},
+      primaryClass={physics.flu-dyn},
+      url={https://arxiv.org/abs/2306.10174}
 }
 
 @dataset{HASSAN2023,
