### 2023-07-11 #5649 @editorialbot
[REVIEW]: BoxKit: A Python library to manage analysis of block-structured simulation datasets
**Submitting author:** @akashdhruv (<a href="http://orcid.org/0000-0003-4997-321X">Akash Dhruv</a>)
**Repository:** https://github.com/akashdhruv/BoxKit
**Branch with paper.md** (empty if default branch): development
**Version:** 2023.12
