+++ boxkit/__meta__.py
@@ -1,8 +1,8 @@
 """Metadata for BoxKit package"""
 
 __pkgname__ = "BoxKit"
-__version__ = "2023.06.7"
+__version__ = "2023.07"
 __authors__ = "Akash Dhruv"
-__license__ = "Apache License"
-__copyright__ = "Copyright (c) Akash Dhruv 2023. All Rights Reserved."
+__license__ = "Apache Software License"
+__copyright__ = "Copyright (c) 2023 UChicago Argonne LLC and contributors. All Rights Reserved."
 __description__ = "BoxKit utility"
