+++ setup.py
@@ -58,7 +58,7 @@ setup(
     long_description=long_description,
     classifiers=[
         "Programming Language :: Python :: 3.8",
-        "License :: OSI Approved :: MIT License",
+        "License :: OSI Approved :: Apache Software License",
     ],
     install_requires=DEPENDENCIES,
     cmdclass={
