+++ .github/workflows/draft-pdf.yml
@@ -2,7 +2,7 @@ on:
   push:
     branches: 
       - main
-      - joss-paper
+      - development
 
 jobs:
   paper:
