### 2023-07-12 #5649 @rvg296
> @rvg296 I believe the command is:
>
> @editorialbot generate my checklist

Yes I have corrected it
