+++ paper/paper.bib
@@ -71,3 +71,8 @@ archivePrefix = "arXiv",
   year = 2023,
   howpublished = {\url{https://www.olcf.ornl.gov/summit/}},
 }
+
+@misc{boxkit-performance,
+  title={{BoxKit Performance}},
+  howpublished= {\url{https://github.com/akashdhruv/boxkit-performance}},
+}
