+++ tests/README.rst
@@ -1,0 +1,2 @@
+Organization of tests based on sites and functionality. All tests should
+be resolved on `development` branch before merge to main.
