### 2023-08-31 #5649 @rvg296
@akashdhruv May be a minor edit, but in line 54 of the paper should the Scikit method be named as `skimage.measure` instead of `skimage_measure `for measuring the bubble shape and size??
