### 2023-09-06 #5649 @akashdhruv
> @akashdhruv May be a minor edit, but in line 54 of the paper should'nt the Scikit method be named as `skimage.measure` instead of `skimage_measure `for measuring the bubble shape and size??

I will address this.
