### 2023-09-27 #5649 @akashdhruv
> @akashdhruv Tried installing this in a virtual environment created with Conda with Py310. I get this installation error. (py310) C:\Users\MENDR>pip install BoxKit --user
>
> ```
> Collecting BoxKit
>   Downloading BoxKit-2023.6.7.tar.gz (852 kB)
>      ---------------------------------------- 852.5/852.5 kB 1.9 MB/s eta 0:00:00
>   Preparing metadata (setup.py) ... error
>   error: subprocess-exited-with-error
>
>   × python setup.py egg_info did not run successfully.
>   │ exit code: 1
>   ╰─> [10 lines of output]
>       Traceback (most recent call last):
>         File "<string>", line 2, in <module>
>         File "<pip-setuptools-caller>", line 34, in <module>
>         File "C:\Users\MENDR\AppData\Local\Temp\1\pip-install-jv7ymwcp\boxkit_f17c2e8a1aef4428bab1ef1a7d2994e5\setup.py", line 14, in <module>
>           import bin.cmd as bin_cmd  # pylint: disable=wrong-import-position
>         File "C:\Users\MENDR\AppData\Local\Temp\1\pip-install-jv7ymwcp\boxkit_f17c2e8a1aef4428bab1ef1a7d2994e5\bin\cmd.py", line 13, in <module>
>           from cbox import cbox_build  # pylint: disable=wrong-import-position
>         File "C:\Users\MENDR\AppData\Local\Temp\1\pip-install-jv7ymwcp\boxkit_f17c2e8a1aef4428bab1ef1a7d2994e5\bin\cbox.py", line 21, in <module>
>           if os.path.exists(os.getenv("PWD") + "/boxkit/depends/boost")
>       TypeError: unsupported operand type(s) for +: 'NoneType' and 'str'
>       [end of output]
>
>   note: This error originates from a subprocess, and is likely not a problem with pip.
> error: metadata-generation-failed
>
> × Encountered error while generating package metadata.
> ╰─> See above for output.
>
> note: This is an issue with the package mentioned above, not pip.
> hint: See above for details.
>
> (py310) C:\Users\MENDR>pip install BoxKit
> Collecting BoxKit
>   Using cached BoxKit-2023.6.7.tar.gz (852 kB)
>   Preparing metadata (setup.py) ... error
>   error: subprocess-exited-with-error
>
>   × python setup.py egg_info did not run successfully.
>   │ exit code: 1
>   ╰─> [10 lines of output]
>       Traceback (most recent call last):
>         File "<string>", line 2, in <module>
>         File "<pip-setuptools-caller>", line 34, in <module>
>         File "C:\Users\MENDR\AppData\Local\Temp\1\pip-install-r2sw90yv\boxkit_decd8282603049f5bb2d36ea2947c869\setup.py", line 14, in <module>
>           import bin.cmd as bin_cmd  # pylint: disable=wrong-import-position
>         File "C:\Users\MENDR\AppData\Local\Temp\1\pip-install-r2sw90yv\boxkit_decd8282603049f5bb2d36ea2947c869\bin\cmd.py", line 13, in <module>
>           from cbox import cbox_build  # pylint: disable=wrong-import-position
>         File "C:\Users\MENDR\AppData\Local\Temp\1\pip-install-r2sw90yv\boxkit_decd8282603049f5bb2d36ea2947c869\bin\cbox.py", line 21, in <module>
>           if os.path.exists(os.getenv("PWD") + "/boxkit/depends/boost")
>       TypeError: unsupported operand type(s) for +: 'NoneType' and 'str'
>       [end of output]
>
>   note: This error originates from a subprocess, and is likely not a problem with pip.
> error: metadata-generation-failed
>
> × Encountered error while generating package metadata.
> ╰─> See above for output.
>
> note: This is an issue with the package mentioned above, not pip.
> hint: See above for details.
> ```

I think this is maybe originating from the command `os.getenv("PWD")` being executed on a Windows Operating System. `$PWD` only exists on Linux/Unix based Operating Systems. I will write a fix for this.

Thank you for reporting this bug.
