### 2023-09-28 #5649 @rvg296
Yes, I have tried. Here's what I get and I believe its related to path

`Collecting boxkit==2023.9
  Downloading BoxKit-2023.9.0.tar.gz (1.4 MB)
     ---------------------------------------- 1.4/1.4 MB 3.2 MB/s eta 0:00:00
  Preparing metadata (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py egg_info did not run successfully.
  │ exit code: 1
  ╰─> [10 lines of output]
      Traceback (most recent call last):
        File "<string>", line 2, in <module>
        File "<pip-setuptools-caller>", line 34, in <module>
        File "C:\Users\MENDR\AppData\Local\Temp\1\pip-install-zydhqfqy\boxkit_9dbb73bbaeb0445ca87594bacf0d2c0f\setup.py", line 49, in <module>
          packages=find_packages(where="./"),
        File "C:\Users\MENDR\AppData\Local\ESRI\conda\envs\py310\lib\site-packages\setuptools\discovery.py", line 127, in find
          convert_path(str(where)),
        File "C:\Users\MENDR\AppData\Local\ESRI\conda\envs\py310\lib\site-packages\setuptools\_distutils\util.py", line 141, in convert_path
          raise ValueError("path '%s' cannot end with '/'" % pathname)
      ValueError: path './' cannot end with '/'
      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.`
