+++ .github/workflows/minimal.yml
@@ -15,8 +15,8 @@ on:
 
 jobs:
 
-  windows:
-    name: "windows"
+  windows-platform:
+    name: "windows-platform"
     runs-on: windows-latest
     - uses: actions/checkout@v2      
     - uses: akashdhruv/Maple/.github/actions/maple@main
@@ -33,8 +33,8 @@ jobs:
           maple container rinse
           maple image delete boxkit
 
-  ubuntu: 
-    name: "ubuntu" 
+  ubuntu-platform: 
+    name: "ubuntu-platform" 
     runs-on: ubuntu-latest
     steps:
     #- name: Login to Docker Hub
