+++ .github/workflows/minimal.yml
@@ -15,9 +15,10 @@ on:
 
 jobs:
 
-  windows-platform:
-    name: "windows-platform"
+  windows:
+    name: "windows"
     runs-on: windows-latest
+    steps:
     - uses: actions/checkout@v2      
     - uses: akashdhruv/Maple/.github/actions/maple@main
     - name: Build Container
@@ -33,8 +34,8 @@ jobs:
           maple container rinse
           maple image delete boxkit
 
-  ubuntu-platform: 
-    name: "ubuntu-platform" 
+  ubuntu: 
+    name: "ubuntu" 
     runs-on: ubuntu-latest
     steps:
     #- name: Login to Docker Hub
