+++ .github/workflows/minimal.yml
@@ -24,10 +24,15 @@ jobs:
     - uses: actions/setup-python@v4
       with:
         python-version: '3.11'
-    - name: Run BoxKit tests
+    - name: Install dependencies
       run: |
           pip3 install click
-          ./setup develop --enable-testing
+          pip3 install -r requirements/core.txt
+    - name: Run setup
+      run: |
+          python3 setup.py develop --user
+    - name: Run tests
+      run: |
           python3 tests/container/create.py
 
   ubuntu: 
