### 2023-09-28 29d3472
update current directory paths
