### 2023-09-28 #5649 @akashdhruv
It looks like I will have to fix the source code to be able to work with either windows or linux machines. I originally intended BoxKit to be used on HPC clusters which are largely linux based.

I have setup a minimal test to replicate your windows test problem: https://github.com/Box-Tools/BoxKit/actions/runs/6343640315
