### 2023-09-28 d4e5960
fix setup for windows
