+++ .github/workflows/minimal.yml
@@ -30,7 +30,7 @@ jobs:
           pip3 install -r requirements/core.txt
     - name: Run setup
       run: |
-          python3 setup.py develop --user
+          python3 setup.py install --user
     - name: Run tests
       run: |
           python3 tests/container/create.py
