+++ boxkit/__meta__.py
@@ -1,7 +1,7 @@
 """Metadata for BoxKit package"""
 
 __pkgname__ = "BoxKit"
-__version__ = "2023.07"
+__version__ = "2023.9.1"
 __authors__ = "Akash Dhruv"
 __license__ = "Apache Software License"
 __copyright__ = "Copyright (c) 2023 UChicago Argonne LLC and contributors. All Rights Reserved."
