### 2023-09-28 #5649 @akashdhruv
@rvg296 It has been fixed. See: https://github.com/Box-Tools/BoxKit/actions/runs/6344172628/job/17233705158?pr=144

Can you try, `pip install boxkit==2023.9.1 --user`

and also, `git+ssh://git@github.com/akashdhruv/BoxKit.git@main --user`
