+++ boxkit/options.py
@@ -1,0 +1,6 @@
+CBOX=None
+PYARROW=None
+ZARR=None
+DASK=None
+SERVER=None
+TESTING=None
