### 2023-10-03 #5649 @rvg296
@akashdhruv , In the paper, you have mentioned the performance is improved ~5x in the paper for measuring the bubble shape and size for Flash-X boiling simulations. Could you please locate that in the box-kit performance repo. Is that in the Jupyter notebook?
