### 2023-10-03 #5649 @akashdhruv
> @akashdhruv , In the paper, you have mentioned the performance is improved ~5x for measuring the bubble shape and size for Flash-X boiling simulations. Could you please locate that in the box-kit performance repo. Is that in the Jupyter notebook?

All the raw data with performance measurements is in the `archive/` The Jupyter notebook contains data picked from the archive.
