+++ boxkit/api/_regionprops.py
@@ -59,12 +59,16 @@ def skimage_props_blk(block, lsetkey, labelkey):
     listprops = skimage_measure.regionprops(block[labelkey].astype(int))
 
     # proplist = ["area", "centroid", "equivalent_diameter_area"]
-    proplist = ["area"]
+    # proplist = ["area"]
 
-    # listprops = [
-    #    {"area": props["area"] * block.dx * block.dy * block.dz} for props in listprops
-    # ]
+    listprops = [
+        {
+            "area": props["area"] * block.dx * block.dy * block.dz,
+            "centroid": props["centroid"],
+        }
+        for props in listprops
+    ]
 
-    listprops = [{key: props[key] for key in proplist} for props in listprops]
+    # listprops = [{key: props[key] for key in proplist} for props in listprops]
 
     return listprops
