### 2023-11-15 bbe483f
Updates to regionprops and resources
