+++ boxkit/resources/flash/_lset.py
@@ -1,0 +1,189 @@
+"""Module for level-set related operations"""
+
+import numpy
+import boxkit
+import skimage.measure as skimage_measure
+
+
+def bubble_contour_plot(ax, dataset, filled=False):
+    """
+    Plot bubble from a dataset to a figure
+
+    Arguments
+    ---------
+    ax         : Axes handle
+    dataset    : Dataset object
+    """
+    merged_dataset = boxkit.mergeblocks(dataset, "dfun", nthreads=1, backend="loky")
+    for block in merged_dataset.blocklist:
+        xmesh, ymesh = numpy.meshgrid(block.xrange("center"), block.yrange("center"))
+
+        if filled:
+            ax.contourf(
+                xmesh[
+                    block.yguard : block.nyb + block.yguard,
+                    block.xguard : block.nxb + block.xguard,
+                ],
+                ymesh[
+                    block.yguard : block.nyb + block.yguard,
+                    block.xguard : block.nxb + block.xguard,
+                ],
+                block["dfun"][
+                    0,
+                    block.yguard : block.nyb + block.yguard,
+                    block.xguard : block.nxb + block.xguard,
+                ],
+            )
+
+        else:
+            ax.contour(
+                xmesh[
+                    block.yguard : block.nyb + block.yguard,
+                    block.xguard : block.nxb + block.xguard,
+                ],
+                ymesh[
+                    block.yguard : block.nyb + block.yguard,
+                    block.xguard : block.nxb + block.xguard,
+                ],
+                block["dfun"][
+                    0,
+                    block.yguard : block.nyb + block.yguard,
+                    block.xguard : block.nxb + block.xguard,
+                ],
+                levels=[0],
+            )
+
+
+def bubble_normal_vectors(ax, merged_dataset, *args, **kwargs):
+    """
+    Plot normal vectors to bubble from a dataset to a figure
+
+    Arguments
+    ---------
+    ax         : Axes handle
+    dataset    : Dataset object
+    """
+    for block in merged_dataset.blocklist:
+        xmesh, ymesh = numpy.meshgrid(block.xrange("center"), block.yrange("center"))
+        ax.contour(
+            xmesh[
+                block.yguard : block.nyb + block.yguard,
+                block.xguard : block.nxb + block.xguard,
+            ],
+            ymesh[
+                block.yguard : block.nyb + block.yguard,
+                block.xguard : block.nxb + block.xguard,
+            ],
+            block["dfun"][
+                0,
+                block.yguard : block.nyb + block.yguard,
+                block.xguard : block.nxb + block.xguard,
+            ],
+            levels=[0],
+        )
+        ax.quiver(
+            xmesh[::5, ::5],
+            ymesh[::5, ::5],
+            block["normx"][
+                0,
+                block.yguard : block.nyb + block.yguard : 5,
+                block.xguard : block.nxb + block.xguard : 5,
+            ],
+            block["normy"][
+                0,
+                block.yguard : block.nyb + block.yguard : 5,
+                block.xguard : block.nxb + block.xguard : 5,
+            ],
+            *args,
+            **kwargs,
+        )
+
+
+def bubble_shape_measurement(dataset):
+    """
+    Perform measurements on the bubble
+    """
+    merged_varlist = ["dfun", "velx", "vely"]
+    merged_dataset = boxkit.mergeblocks(
+        dataset, merged_varlist, nthreads=1, backend="loky"
+    )
+    merged_dataset.fill_guard_cells(merged_varlist)
+
+    bubblelist = boxkit.regionprops(merged_dataset, "dfun", backend="loky", nthreads=1)
+
[…]
