+++ boxkit/resources/flash/_lset.py
@@ -5,7 +5,7 @@ import boxkit
 import skimage.measure as skimage_measure
 
 
-def bubble_contour_plot(ax, dataset, filled=False):
+def lset_plot_contour_2d(ax, merged_dataset, filled=False, *args, **kwargs):
     """
     Plot bubble from a dataset to a figure
 
@@ -14,7 +14,11 @@ def bubble_contour_plot(ax, dataset, filled=False):
     ax         : Axes handle
     dataset    : Dataset object
     """
-    merged_dataset = boxkit.mergeblocks(dataset, "dfun", nthreads=1, backend="loky")
+    if len(merged_dataset.blocklist) > 1:
+        raise ValueError(
+            "[boxkit.resources.flash.lset_contour_plot_2d] dataset must only have one block"
+        )
+
     for block in merged_dataset.blocklist:
         xmesh, ymesh = numpy.meshgrid(block.xrange("center"), block.yrange("center"))
 
@@ -33,6 +37,8 @@ def bubble_contour_plot(ax, dataset, filled=False):
                     block.yguard : block.nyb + block.yguard,
                     block.xguard : block.nxb + block.xguard,
                 ],
+                *args,
+                **kwargs,
             )
 
         else:
@@ -51,10 +57,12 @@ def bubble_contour_plot(ax, dataset, filled=False):
                     block.xguard : block.nxb + block.xguard,
                 ],
                 levels=[0],
+                *args,
+                **kwargs,
             )
 
 
-def bubble_normal_vectors(ax, merged_dataset, *args, **kwargs):
+def lset_plot_normals_2d(ax, merged_dataset, *args, **kwargs):
     """
     Plot normal vectors to bubble from a dataset to a figure
 
@@ -63,27 +71,28 @@ def bubble_normal_vectors(ax, merged_dataset, *args, **kwargs):
     ax         : Axes handle
     dataset    : Dataset object
     """
+    if len(merged_dataset.blocklist) > 1:
+        raise ValueError(
+            "[boxkit.resources.flash.lset_normal_vectors_2d] dataset must only have one block"
+        )
+
+    varlist = ["normx", "normy"]
+    for ivar in varlist:
+        merged_dataset.addvar(ivar, dtype=float)
+
+    lset_compute_normals_2d(merged_dataset, varlist)
+
     for block in merged_dataset.blocklist:
         xmesh, ymesh = numpy.meshgrid(block.xrange("center"), block.yrange("center"))
-        ax.contour(
+        ax.quiver(
             xmesh[
-                block.yguard : block.nyb + block.yguard,
-                block.xguard : block.nxb + block.xguard,
+                block.yguard : block.nyb + block.yguard : 5,
+                block.xguard : block.nxb + block.xguard : 5,
             ],
             ymesh[
-                block.yguard : block.nyb + block.yguard,
-                block.xguard : block.nxb + block.xguard,
+                block.yguard : block.nyb + block.yguard : 5,
+                block.xguard : block.nxb + block.xguard : 5,
             ],
-            block["dfun"][
-                0,
-                block.yguard : block.nyb + block.yguard,
-                block.xguard : block.nxb + block.xguard,
-            ],
-            levels=[0],
-        )
-        ax.quiver(
-            xmesh[::5, ::5],
-            ymesh[::5, ::5],
             block["normx"][
                 0,
                 block.yguard : block.nyb + block.yguard : 5,
@@ -98,34 +107,16 @@ def bubble_normal_vectors(ax, merged_dataset, *args, **kwargs):
             **kwargs,
         )
 
+    for var in varlist:
+        merged_dataset.delvar(var)
 
-def bubble_shape_measurement(dataset):
+
+def lset_compute_normals_2d(dataset, varlist):
     """
-    Perform measurements on the bubble
[…]
