+++ .github/workflows/linting.yml
@@ -38,7 +38,7 @@ jobs:
           sudo apt-get install -y python3 python3-dev python3-pip
           sudo apt-get install -y python-is-python3
           pip3 install black pylint
-          pip3 install -r requirements/core.txt -r requirements/cbox.txt -r requirements/dask.txt -r requirements/pyarrow.txt -r requirements/server.txt -r requirements/testing.txt -r requirements/zarr.txt
+          pip3 install -r requirements/core.txt -r requirements/cbox.txt -r requirements/dask.txt -r requirements/pyarrow.txt -r requirements/server.txt -r requirements/analysis.txt -r requirements/zarr.txt
     - name: Run linting 
       run: |
           pylint boxkit/library/*.py
@@ -53,7 +53,7 @@ jobs:
           sudo apt-get install -y python3 python3-dev python3-pip
           sudo apt-get install -y python-is-python3
           pip3 install black pylint
-          pip3 install -r requirements/core.txt -r requirements/cbox.txt -r requirements/dask.txt -r requirements/pyarrow.txt -r requirements/server.txt -r requirements/testing.txt -r requirements/zarr.txt
+          pip3 install -r requirements/core.txt -r requirements/cbox.txt -r requirements/dask.txt -r requirements/pyarrow.txt -r requirements/server.txt -r requirements/analysis.txt -r requirements/zarr.txt
     - name: Run linting 
       run: |
           pylint boxkit/api/*.py
@@ -68,7 +68,7 @@ jobs:
           sudo apt-get install -y python3 python3-dev python3-pip
           sudo apt-get install -y python-is-python3
           pip3 install black pylint
-          pip3 install -r requirements/core.txt -r requirements/cbox.txt -r requirements/dask.txt -r requirements/pyarrow.txt -r requirements/server.txt -r requirements/testing.txt -r requirements/zarr.txt
+          pip3 install -r requirements/core.txt -r requirements/cbox.txt -r requirements/dask.txt -r requirements/pyarrow.txt -r requirements/server.txt -r requirements/analysis.txt -r requirements/zarr.txt
     - name: Run linting 
       run: |
           pylint boxkit/resources/read/*.py
