+++ boxkit/options.py
@@ -3,4 +3,4 @@ PYARROW=None
 ZARR=None
 DASK=None
 SERVER=None
-ANALYSIS=1
+ANALYSIS=None
