### 2023-11-16 6e95153
add numba optimizations
