+++ boxkit/resources/flash/_lset.py
@@ -3,6 +3,7 @@
 import numpy
 import boxkit
 import skimage.measure as skimage_measure
+from numba import jit
 
 
 def lset_plot_contour_2d(ax, merged_dataset, filled=False, *args, **kwargs):
@@ -136,125 +137,209 @@ def lset_compute_normals_2d(dataset, varlist):
     dataset.fill_guard_cells(varlist)
 
 
-def lset_shape_measurement_2d(merged_dataset):
+def lset_vel_measurement_2d(merged_dataset):
     """
-    Perform measurements on the bubble
+    Perform velocity measurement on level set
+    """
+    if len(merged_dataset.blocklist) > 1:
+        raise ValueError(
+            "[boxkit.resources.flash.lset_vel_measurement_2d] dataset must only have one block"
+        )
+
+    merged_dataset.addvar("bwlabel", dtype=int)
+
+    for block in merged_dataset.blocklist:
+        block["bwlabel"] = skimage_measure.label(block["dfun"] >= 0)
+        mean_vel = numpy.zeros([numpy.max(block["bwlabel"][:]), 2], dtype=float)
+        mean_label = numpy.zeros([numpy.max(block["bwlabel"][:]), 1], dtype=float)
+
+        lset_vel_blk_2d(
+            block["dfun"],
+            block["bwlabel"],
+            block["velx"],
+            block["vely"],
+            block.nxb,
+            block.nyb,
+            block.xguard,
+            block.yguard,
+            block.dx,
+            block.dy,
+            mean_vel,
+            mean_label,
+        )
+
+    mean_vel = mean_vel / mean_label
+
+    merged_dataset.delvar("bwlabel")
+
+    return mean_vel
+
+
+def lset_shape_measurement_2d(merged_dataset, correction=False):
+    """
+    Perform shape measurement on level set
     """
     if len(merged_dataset.blocklist) > 1:
         raise ValueError(
             "[boxkit.resources.flash.lset_shape_measurement_2d] dataset must only have one block"
         )
 
-    bubblelist = boxkit.regionprops(merged_dataset, "dfun", backend="loky", nthreads=1)
+    bubblelist = boxkit.regionprops(merged_dataset, "dfun")
 
-    merged_dataset.addvar("bwlabel", dtype=int)
-    merged_dataset.addvar("nrmx", dtype=float)
-    merged_dataset.addvar("nrmy", dtype=float)
+    if correction:
 
-    for block in merged_dataset.blocklist:
-        block["bwlabel"] = skimage_measure.label(block["dfun"] >= 0)
+        merged_dataset.addvar("bwlabel", dtype=int)
+        merged_dataset.addvar("nrmx", dtype=float)
+        merged_dataset.addvar("nrmy", dtype=float)
 
-    lset_compute_normals_2d(merged_dataset, ["nrmx", "nrmy"])
+        lset_compute_normals_2d(merged_dataset, ["nrmx", "nrmy"])
 
-    modified_perimeter = [0.0] * len(bubblelist)
+        modified_perimeter = numpy.zeros([len(bubblelist), 1], dtype=float)
+        sol_points = numpy.zeros([2, 2], dtype=float)
 
-    for block in merged_dataset.blocklist:
+        for block in merged_dataset.blocklist:
+            block["bwlabel"] = skimage_measure.label(block["dfun"] >= 0)
+            xcenter = block.xrange("center")
+            ycenter = block.yrange("center")
 
-        xcenter = block.xrange("center")
-        ycenter = block.yrange("center")
+            lset_perimeter_blk_2d(
+                block["dfun"],
+                block["bwlabel"],
+                block["nrmx"],
+                block["nrmy"],
+                xcenter,
+                ycenter,
+                modified_perimeter,
+                block.nxb,
[…]
