+++ requirements/analysis.txt
@@ -1,1 +1,2 @@
 scikit-image>=0.18.1
+numba
