+++ .github/actions/install-analysis/action.yml
@@ -1,0 +1,23 @@
+# Name of the composite action
+name: install
+
+inputs:
+  image: 
+    description: 'Input image for boxkit'
+    required: true
+
+runs:
+  using: "composite"
+  steps:
+    - run: maple container pour --image=${{inputs.image}} 
+      shell: bash
+    - run: maple container execute "pip3 install click"
+      shell: bash
+    - run: maple container execute "export CXX=g++ BOOST_LIB_DIR=usr/lib BOOST_INCLUDE_DIR=/usr/include && ./setup install --with-cbox"
+      shell: bash
+    - run: maple container execute "./setup clean"
+      shell: bash
+    - run: maple container commit --image=${{inputs.image}}
+      shell: bash
+    - run: maple container rinse
+      shell: bash
