### 2023-11-17 a0d8bbf
--with-analysis -> --enable-analysis
