### 2023-12-04 #5649 @kellyrowland
@akashdhruv can you please:

- make a release of the software with the latest changes from the review and post the version number here
- archive the release on Zenodo, figshare, etc. and post the DOI here
