### 2023-12-04 #5649 @akashdhruv
> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
> * [ ]  Double check authors and affiliations (including ORCIDs)
> * [ ]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * [ ]  Archive the release on Zenodo/figshare/etc and post the DOI here.
> * [ ]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * [ ]  Make sure that the license listed for the archive is the same as the software license.

I cannot make changes to the comment above so I am copying the checklist here:
