### 2023-12-04 #5649 @kellyrowland
@akashdhruv would you be able to change the archive title from "Box-Tools/BoxKit: Peer reviewed release for JOSS" to "BoxKit: A Python library to manage analysis of block-structured simulation datasets" to match that of the submission?
