### 2023-12-04 #5649 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
